HHH-14006 Test extended bytecode enhancement for mapped superclasses and entity superclasses
