+++ hibernate-core/src/test/java/org/hibernate/test/bytecode/enhancement/basic/BasicEnhancementTest.java
@@ -8,7 +8,7 @@ package org.hibernate.test.bytecode.enhancement.basic;
 
 import org.hibernate.engine.spi.ManagedEntity;
 import org.hibernate.engine.spi.PersistentAttributeInterceptable;
-import org.hibernate.engine.spi.PersistentAttributeInterceptor;
+
 import org.hibernate.testing.bytecode.enhancement.BytecodeEnhancerRunner;
 import org.hibernate.testing.bytecode.enhancement.EnhancerTestUtils;
 import org.junit.Assert;
@@ -20,7 +20,6 @@ import javax.persistence.Id;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Set;
 
 import static org.hibernate.testing.junit4.ExtraAssertions.assertTyping;
 import static org.junit.Assert.assertArrayEquals;
@@ -111,104 +110,6 @@ public class BasicEnhancementTest {
 
         entity.someStrings = Arrays.asList( "A", "B", "C", "D" );
         assertArrayEquals( new String[]{"A", "B", "C", "D"}, entity.someStrings.toArray() );
-    }
-
-    // --- //
-
-    public static class ObjectAttributeMarkerInterceptor implements PersistentAttributeInterceptor {
-
-        public static final Object READ_MARKER = new Object();
-        public static final Object WRITE_MARKER = new Object();
-
-        @Override
-        public boolean readBoolean(Object obj, String name, boolean oldValue) {
-            return oldValue;
-        }
-
-        @Override
-        public boolean writeBoolean(Object obj, String name, boolean oldValue, boolean newValue) {
-            return newValue;
-        }
-
-        @Override
-        public byte readByte(Object obj, String name, byte oldValue) {
-            return oldValue;
-        }
-
-        @Override
-        public byte writeByte(Object obj, String name, byte oldValue, byte newValue) {
-            return newValue;
-        }
-
-        @Override
-        public char readChar(Object obj, String name, char oldValue) {
-            return oldValue;
-        }
-
-        @Override
-        public char writeChar(Object obj, String name, char oldValue, char newValue) {
-            return newValue;
-        }
-
-        @Override
-        public short readShort(Object obj, String name, short oldValue) {
-            return oldValue;
-        }
-
-        @Override
-        public short writeShort(Object obj, String name, short oldValue, short newValue) {
-            return newValue;
-        }
-
-        @Override
-        public int readInt(Object obj, String name, int oldValue) {
-            return oldValue;
-        }
-
-        @Override
-        public int writeInt(Object obj, String name, int oldValue, int newValue) {
-            return newValue;
-        }
-
-        @Override
-        public float readFloat(Object obj, String name, float oldValue) {
-            return oldValue;
-        }
-
-        @Override
-        public float writeFloat(Object obj, String name, float oldValue, float newValue) {
-            return newValue;
-        }
-
-        @Override
-        public double readDouble(Object obj, String name, double oldValue) {
-            return oldValue;
-        }
-
-        @Override
-        public double writeDouble(Object obj, String name, double oldValue, double newValue) {
-            return newValue;
-        }
-
-        @Override
-        public long readLong(Object obj, String name, long oldValue) {
-            return oldValue;
-        }
-
-        @Override
-        public long writeLong(Object obj, String name, long oldValue, long newValue) {
-            return newValue;
-        }
-
-        @Override
-        public Object readObject(Object obj, String name, Object oldValue) {
-            return READ_MARKER;
-        }
-
-        @Override
-        public Object writeObject(Object obj, String name, Object oldValue, Object newValue) {
-            return WRITE_MARKER;
-        }
     }
 
     // --- //
