+++ hibernate-core/src/test/java/org/hibernate/test/bytecode/enhancement/basic/MappedSuperclassTest.java
@@ -2,10 +2,14 @@ package org.hibernate.test.bytecode.enhancement.basic;
 
 import org.hibernate.bytecode.enhance.spi.DefaultEnhancementContext;
 import org.hibernate.bytecode.enhance.spi.UnloadedClass;
+import org.hibernate.engine.spi.PersistentAttributeInterceptable;
+
 import org.hibernate.testing.TestForIssue;
 import org.hibernate.testing.bytecode.enhancement.BytecodeEnhancerRunner;
 import org.hibernate.testing.bytecode.enhancement.CustomEnhancementContext;
 import org.hibernate.testing.bytecode.enhancement.EnhancerTestContext;
+import org.hibernate.testing.bytecode.enhancement.EnhancerTestUtils;
+import org.junit.Assume;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -16,6 +20,9 @@ import javax.persistence.Version;
 
 import static org.hibernate.testing.bytecode.enhancement.EnhancerTestUtils.checkDirtyTracking;
 import static org.hibernate.testing.bytecode.enhancement.EnhancerTestUtils.clearDirtyTracking;
+import static org.junit.Assert.assertNotSame;
+import static org.junit.Assert.assertSame;
+import static org.junit.Assume.assumeTrue;
 
 /**
  * @author Luis Barreiro
@@ -41,10 +48,38 @@ public class MappedSuperclassTest {
         checkDirtyTracking( charles, "title", "oca" );
     }
 
+    // Adapted from BasicEnhancementTest#basicExtendedEnhancementTest
+    @Test
+    @TestForIssue(jiraKey = "HHH-14006")
+    public void extendedEnhancementTest() {
+        // This test only works if lazy loading bytecode enhancement is enabled,
+        // otherwise extended bytecode enhancement does not do anything we can check.
+        assumeTrue( PersistentAttributeInterceptable.class.isAssignableFrom( Employee.class ) );
+
+        Employee entity = new Employee();
+        ( (PersistentAttributeInterceptable) entity ).$$_hibernate_setInterceptor( new ObjectAttributeMarkerInterceptor() );
+
+        Object decoy = new Object();
+        // This accesses "name" on a variable of type Employee,
+        // but "anUnspecifiedObject" is defined the superclass Person.
+        // Such "virtual" access used to break extended bytecode enhancement.
+        entity.anUnspecifiedObject = decoy;
+
+        Object gotByReflection = EnhancerTestUtils.getFieldByReflection( entity, "anUnspecifiedObject" );
+        assertNotSame( decoy, gotByReflection );
+        assertSame( ObjectAttributeMarkerInterceptor.WRITE_MARKER, gotByReflection );
+
+        Object gotByEnhancedDirectAccess = entity.anUnspecifiedObject;
+        assertNotSame( decoy, gotByEnhancedDirectAccess );
+        assertSame( ObjectAttributeMarkerInterceptor.READ_MARKER, gotByEnhancedDirectAccess );
+    }
+
     // --- //
 
     @MappedSuperclass
     private static class Person {
+
+        Object anUnspecifiedObject;
 
         @Id
         String name;
